pyproject.toml
[tool.poetry]
name = "genai-takehome"
version = "0.1.0"

authors = ["pablo.sanjose <pablo.sanjose.villar@gmail.com>"]
description = ""
readme = "README.md"

packages = [
  {include = "app"},
]

[tool.poetry.dependencies]
python = "^3.10"

chromadb = "==0.5.3"
httpx = "^0.27.0"
langchain-chroma = "^0.1.2"
langchain-community = "*"
langchain-core = "*"
langchain-openai = "*"
langserve = {extras = ["all"], version = "^0.2.2"}
pandas = "^2.2.2"
pydantic = "<2"
sqlalchemy = "*"
python-dotenv = "^1.0.1"

[tool.poetry.group.ui.dependencies]
httpx = "^0.27.0"
httpx-sse = "~=0.4.0"
streamlit = "==1.35.0"

[tool.poetry.group.dev.dependencies]
fastapi-cli = "^0.0.5"
ruff = "^0.5.6"
watchdog = "^4.0.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry.scripts]
netflix-data = 'app.tools.netflix:main'
